requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
python-whois>=0.8.0
rich>=13.7.0
click>=8.1.0
jinja2>=3.1.0
urllib3>=2.1.0
cryptography>=41.0.0
